AL-1857 Fix Composite Monitors Cannot Use Dry Run
Fixed issue where if one tried to dry-run creating a new composite
monitor, it wouldn't work because the sub-monitors would not be
returned. Updated the helper function that actually creates the monitors
to always return a valid monitor configuration. Additionally, if it is
creating a new monitor but is in dry-run mode, the helper function now
generates a fake monitor id and includes that in the response. Also
added an unit test covering this behavior.