AL-1944 Fix Interpolation Bug
There was an issue where the validation for the config would not allow
variables for certain options in the `option` section. Created a new
Regex validator that looks for the pattern of a variable and allows
that.

Also updated a unit test to verify that variables in subsections are
interpolated.